all plugin config moved to plugins/config
