+++ mo-vim/.config/nvim/lua/core/init.lua
@@ -1,3 +1,4 @@
 require "core.bootstrap"
 require "core.options"
 require "core.autocommands"
+require "core.ftdetect"
