+++ mo-vim/.config/nvim/lua/plugins/config/nvim-tree.lua
