+++ mo-vim/.config/nvim/lua/plugins/config/treesitter.lua
@@ -1,7 +1,7 @@
-local status_ok, treesitter = pcall(require, "nvim-treesitter")
-if not status_ok then
-	return
-end
+-- local status_ok, treesitter = pcall(require, "nvim-treesitter")
+-- if not status_ok then
+-- 	return
+-- end
 
 local status_ok, configs = pcall(require, "nvim-treesitter.configs")
 if not status_ok then
