+++ mo-vim/.config/nvim/lua/plugins/init.lua
@@ -1,19 +1,8 @@
 require "plugins.lazy"
 require "plugins.config.keymaps"
--- require "plugins.cmp"
-require "plugins.telescope"
-require "plugins.treesitter"
-require "plugins.comment"
-require "plugins.nvim-tree"
-require "plugins.bufferline"
-require "plugins.lualine"
-require "plugins.toggleterm"
-require "plugins.project"
-require "plugins.illuminate"
-require "plugins.indentline"
+require "plugins.config.bufferline"
+require "plugins.config.nvim-tree"
+require "plugins.config.lualine"
+require "plugins.config.illuminate"
 require "plugins.lsp"
-require "plugins.zen"
-require "plugins.snippets"
-require "plugins.ftdetect"
-require "plugins.leap"
 require "plugins.config.harpoon"
